Fixed DateType issue
