+++ UBL/UBL.csproj
@@ -5,7 +5,7 @@
     <Nullable>enable</Nullable>
     <AssemblyName>Hilma.UBL</AssemblyName>
     <RootNamespace>Hilma.UBL</RootNamespace>
-    <Version>0.1.1</Version>
+    <Version>0.1.2</Version>
     <Authors>Hankintailmoitukset.fi</Authors>
     <Product>Hilma UBL</Product>
     <Description>Open project for EU procurements related UBL data model</Description>
