+++ UBL/UnqualifiedDataTypes/DateType.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.ComponentModel.DataAnnotations;
 
 namespace UBL.UnqualifiedDataTypes
 {
@@ -8,6 +7,8 @@ namespace UBL.UnqualifiedDataTypes
   /// </summary>
   public class DateType
   {
+    private DateTime _value;
+
     public DateType()
     {
     }
@@ -19,8 +20,8 @@ namespace UBL.UnqualifiedDataTypes
 
     public DateTime Value
     {
-      get => Value;
-      set => value = new DateTime(value.Year, value.Month, value.Day);
+      get => _value;
+      set => _value = new DateTime(value.Year, value.Month, value.Day);
     }
 
     public static implicit operator DateTime(DateType d) => d.Value;
@@ -29,38 +30,6 @@ namespace UBL.UnqualifiedDataTypes
     public override string ToString()
     {
       return $"{Value:CCYY-MM-DD}";
-    }
-  }
-
-  /// <summary>
-  /// An instance of time that occurs every day.
-  /// </summary>
-  public class TimeType
-  {
-    public TimeType(int hour, int minute, int second)
-    {
-      Hour = hour;
-      Minute = minute;
-      Second = second;
-    }
-
-    public TimeType()
-    {
-    }
-
-    [Range(0,24)]
-    public int Hour { get; set; }
-    [Range(0, 60)]
-    public int Minute { get; set; }
-
-    [Range(0, 60)]
-    public int Second { get; set; }
-    
-    public static implicit operator TimeType(DateTime d) => new TimeType(d.Hour, d.Minute, d.Second );
-
-    public override string ToString()
-    {
-      return $"{Hour:00}:{Minute:00}:{Second:00}";
     }
   }
 }
